+++ CHANGELOG.md
@@ -2,11 +2,16 @@
 
 ## Unreleased
 
+-Features for next release
+
 ### Misc. Updates
 
 - Model updates for the following services
   - AWSLocation
   - AWSPolly
+
+- **AWSCore**
+  - Improves `AWSSynchronizedMutableDictionary` and adds unit tests (See [PR #4051](https://github.com/aws-amplify/aws-sdk-ios/pull/4051))
 
 ## 2.27.4